Domain decomposition utils module created
