+++ examples/paper_tests/test5_domain_decomposition/sources_grouping_demo.py
@@ -1,175 +1,10 @@
 import math
 import importlib
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import List, Tuple, Optional
 import numpy as np
 
-@dataclass
-class Source:
-    """Represents a single radiation source.
-
-    Attributes
-    ----------
-    pos : np.ndarray
-        Source position in 3D Cartesian coordinates (shape `(3,)`).
-    strength : float
-        Source emissivity/intensity used by downstream physics models.
-    radius : float
-        Maximum tracing/influence radius for this source.
-    gid : Optional[int]
-        Optional global source identifier.
-    """
-    pos: np.ndarray          # shape (3,)
-    strength: float          # e.g. photon rate
-    radius: float            # maximum tracing radius
-    gid: Optional[int] = None 
-
-
-@dataclass
-class Group:
-    """Container for a set of sources grouped for joint processing.
-
-    Attributes
-    ----------
-    sources : List[Source]
-        Sources that belong to this group.
-    center : np.ndarray
-        Center of the enclosing sphere of grouped source spheres (shape `(3,)`).
-    radius : float
-        Radius of the enclosing sphere.
-    bbox_min : np.ndarray
-        Minimum corner of the axis-aligned bounding box around the group.
-    bbox_max : np.ndarray
-        Maximum corner of the axis-aligned bounding box around the group.
-    nvox_local : int
-        Estimated number of local voxels covered by the group bounding box.
-    cost : float
-        Estimated computational cost for this group.
-    """
-    sources: List[Source] = field(default_factory=list)
-    center: np.ndarray = field(default_factory=lambda: np.zeros(3))
-    radius: float = 0.0      # enclosing sphere radius for union of balls
-    bbox_min: np.ndarray = field(default_factory=lambda: np.zeros(3))
-    bbox_max: np.ndarray = field(default_factory=lambda: np.zeros(3))
-    nvox_local: int = 0
-    cost: float = 0.0
-
-    def source_ids(self):
-        """Return the list of source global IDs in this group."""
-        return [s.gid for s in self.sources]
-
-
-def enclosing_sphere(
-    centers: np.ndarray,
-    radii: np.ndarray,
-    max_iter: int = 200,
-    tol: float = 1e-8,
-) -> Tuple[np.ndarray, float]:
-    """Approximate the minimum enclosing sphere of spheres/balls.
-
-    The objective is:
-        minimize_c max_i ||c - x_i|| + r_i
-
-    Parameters
-    ----------
-    centers : np.ndarray
-        Sphere centers, shape `(N, 3)`.
-    radii : np.ndarray
-        Sphere radii, shape `(N,)`.
-    max_iter : int
-        Maximum number of fixed-point iterations.
-    tol : float
-        Convergence tolerance on center displacement.
-
-    Returns
-    -------
-    Tuple[np.ndarray, float]
-        Estimated enclosing sphere center and radius.
-    """
-    if len(centers) == 0:
-        return np.zeros(3), 0.0
-    if len(centers) == 1:
-        return centers[0].copy(), float(radii[0])
-
-    c = centers.mean(axis=0)
-
-    for k in range(max_iter):
-        d = np.linalg.norm(centers - c[None, :], axis=1) + radii
-        j = np.argmax(d)
-        direction = centers[j] - c
-        norm = np.linalg.norm(direction)
-        if norm > 0.0:
-            direction = direction / norm
-        else:
-            direction = np.zeros(3)
-
-        # diminishing step
-        eta = 1.0 / (k + 2.0)
-        c_new = c + eta * direction * max(1e-12, np.linalg.norm(centers[j] - c))
-
-        if np.linalg.norm(c_new - c) < tol:
-            c = c_new
-            break
-        c = c_new
-
-    R = np.max(np.linalg.norm(centers - c[None, :], axis=1) + radii)
-    return c, float(R)
-
-
-def bounding_box_from_sphere(center: np.ndarray, radius: float) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Compute axis-aligned bounding box corners from sphere center/radius.
-
-    Parameters
-    ----------
-    center : np.ndarray
-        Sphere center (shape `(3,)`).
-    radius : float
-        Sphere radius.
-
-    Returns
-    -------
-    Tuple[np.ndarray, np.ndarray]
-        Minimum and maximum corners of the bounding box.
-    """
-    return center - radius, center + radius
-
-
-def morton_like_key(p: np.ndarray, domain_min: np.ndarray, domain_max: np.ndarray, bits: int = 10) -> int:
-    """
-    Lightweight Morton-like ordering. 
-    Maps point to integer grid then interleaves bits.
-
-    Parameters
-    ----------
-    p : np.ndarray
-        Point coordinates (shape `(3,)`).
-    domain_min : np.ndarray
-        Minimum corner of the domain (shape `(3,)`).
-    domain_max : np.ndarray
-        Maximum corner of the domain (shape `(3,)`).
-    bits : int
-        Number of bits per dimension for the grid. Total key bits will be 3x this.
-
-    Returns
-    -------
-    int     Morton-like key for the point.
-    """
-    # Normalize to [0, 1]
-    x = np.clip((p - domain_min) / np.maximum(domain_max - domain_min, 1e-12), 0.0, 1.0 - 1e-12)
-
-    # Scale to integer by shifting by bits. 
-    grid = (x * (1 << bits)).astype(int)
-
-    def split_by_3(v: int) -> int:
-        out = 0
-        for i in range(bits):
-            out |= ((v >> i) & 1) << (3 * i)
-        return out
-
-    return split_by_3(grid[0]) | (split_by_3(grid[1]) << 1) | (split_by_3(grid[2]) << 2)
-
-
+import pyc2ray.utils.domain_decomposition_utils as dd_utils
 
 @dataclass
 class VariableResolutionGrid:
@@ -204,141 +39,6 @@ class VariableResolutionGrid:
             if vol > 0.0:
                 total += vol / (dx ** 3)
         return int(math.ceil(total))
-
-
-def evaluate_group(group_sources: List[Source], grid: VariableResolutionGrid) -> Group:
-    """
-    Build a `Group` and compute its geometric and cost properties.
-
-    Parameters
-    ----------
-    group_sources : List[Source]
-        Sources that belong to this group.
-    grid : VariableResolutionGrid
-        Grid model used to estimate local voxel counts.
-
-    Returns
-    -------
-    Group     Group object with computed center, radius, bounding box, local voxel count, and cost.
-    """
-    centers = np.array([s.pos for s in group_sources], dtype=float)
-    radii = np.array([s.radius for s in group_sources], dtype=float)
-
-    c, R = enclosing_sphere(centers, radii)
-    bbox_min, bbox_max = bounding_box_from_sphere(c, R)
-    nvox = grid.estimate_voxels_in_bbox(bbox_min, bbox_max)
-
-    # simple cost model: number of sources times local voxel count
-    cost = len(group_sources) * nvox
-
-    return Group(
-        sources=list(group_sources),
-        center=c,
-        radius=R,
-        bbox_min=bbox_min,
-        bbox_max=bbox_max,
-        nvox_local=nvox,
-        cost=cost,
-    )
-
-
-def build_groups(
-    sources: List[Source],
-    grid: VariableResolutionGrid,
-    r_group_max: float,
-    nsrc_max: int,
-    nvox_max: int,
-    cost_max: float,
-) -> List[Group]:
-    """Grouping of sources in Morton-like spatial order.
-
-    Sources are sorted spatially, then accumulated into a running group until
-    the radius constraint is violated.
-
-    Parameters
-    ----------
-    sources : List[Source]
-        List of sources to group.
-    grid : VariableResolutionGrid
-        Grid model used to estimate local voxel counts for groups.
-    r_group_max : float
-        Maximum allowed group radius (enclosing sphere of source spheres).
-    nsrc_max : int
-        Maximum allowed number of sources in a group.
-    nvox_max : int
-        Maximum allowed number of local voxels covered by the group bounding box.
-    cost_max : float
-        Maximum allowed computational cost for the group.
-
-    Returns
-    -------
-    List[Group]     List of groups that satisfy the constraints.
-    """
-
-    if not sources:
-        return []
-
-    # spatial ordering
-    dmin, dmax = grid.domain_min, grid.domain_max
-    ordered = sorted(sources, key=lambda s: morton_like_key(s.pos, dmin, dmax))
-
-    groups: List[Group] = []
-    current: List[Source] = []
-
-    def valid(g: Group) -> bool:
-        return (
-            g.radius <= r_group_max
-            # and len(g.sources) <= nsrc_max
-            # and g.nvox_local <= nvox_max
-            # and g.cost <= cost_max
-        )
-
-    for s in ordered:
-        if not current:
-            current = [s]
-            continue
-
-        trial = current + [s]
-        gtrial = evaluate_group(trial, grid)
-
-        if valid(gtrial):
-            current = trial
-        else:
-            groups.append(evaluate_group(current, grid))
-            current = [s]
-
-    if current:
-        groups.append(evaluate_group(current, grid))
-
-    return groups
-
-
-def assign_groups_to_ranks(groups: List[Group], nranks: int):
-    """
-    Groups to ranks assignement according to cost.
-
-    Parameters
-    ----------
-    groups : List[Group]
-        List of groups to assign.
-    nranks : int
-        Number of ranks to distribute groups across.
-
-    Returns
-    Tuple[List[List[Group]], List[float]]
-        A tuple of (rank_groups, rank_costs), where rank_groups is a list of lists of groups assigned to each rank,
-        and rank_costs is the total cost for each rank. 
-    """
-    rank_groups = [[] for _ in range(nranks)]
-    rank_costs = [0.0 for _ in range(nranks)]
-
-    for g in sorted(groups, key=lambda x: x.cost, reverse=True):
-        r = int(np.argmin(rank_costs))
-        rank_groups[r].append(g)
-        rank_costs[r] += g.cost
-
-    return rank_groups, rank_costs
-
 
 
 def generate_sources(num_cluster_sources: int = 70, cluster_center: np.ndarray = None, cluster_width: float = 0.02,
@@ -381,13 +81,13 @@ def generate_sources(num_cluster_sources: int = 70, cluster_center: np.ndarray =
     for i in range(num_cluster_sources):
         pos = cluster_center + rng.normal(scale=cluster_sigma, size=3)
         strength = source_strength
-        sources.append(Source(pos=pos, strength=strength, radius=r_max_lls, gid=i))
+        sources.append(dd_utils.Source(pos=pos, strength=strength, radius=r_max_lls, gid=i))
 
     # Sparse sources across the full box
     for i in range(num_cluster_sources, num_cluster_sources + num_sparse_sources):
         pos = rng.uniform(0, boxsize, size=3)
         strength = source_strength
-        sources.append(Source(pos=pos, strength=strength, radius=r_max_lls, gid=i))
+        sources.append(dd_utils.Source(pos=pos, strength=strength, radius=r_max_lls, gid=i))
 
     return sources
 
@@ -475,8 +175,8 @@ def _box_faces(pmin: np.ndarray, pmax: np.ndarray):
 
 def plot_grid_and_sources(
     grid: VariableResolutionGrid,
-    sources: List[Source],
-    groups: Optional[List[Group]] = None,
+    sources: List[dd_utils.Source],
+    groups: Optional[List[dd_utils.Group]] = None,
     plot_sources: bool = True,
     plot_groups: bool = True,
     plot_bbox: bool = True
@@ -666,7 +366,7 @@ def main():
     )
 
     # Build groups with a uniform radius constraint based on the LLS mean free path.
-    groups = build_groups(
+    groups = dd_utils.build_groups(
         sources=sources,
         grid=grid,
         r_group_max=1.5 * r_max_lls,
@@ -676,7 +376,7 @@ def main():
     )
 
     # Distribute groups across ranks by estimated cost.
-    rank_groups, rank_costs = assign_groups_to_ranks(groups, nranks=4)
+    rank_groups, rank_costs = dd_utils.assign_groups_to_ranks(groups, nranks=4)
 
     # Report results and plot. 
     print(f"Toy box size                = {boxsize:.1f} cMpc/h")
@@ -705,7 +405,7 @@ def main():
                 f"nvox={g.nvox_local:7d}, cost={g.cost:10.1f}"
             )
 
-    plot_grid_and_sources(grid, sources, groups, )
+    # plot_grid_and_sources(grid, sources, groups, )
 
 if __name__ == "__main__":
     main()
