+++ pyc2ray/utils/__init__.py
@@ -2,3 +2,4 @@
 from .logutils import *  # noqa: F403
 from .other_utils import *  # noqa: F403
 from .sourceutils import *  # noqa: F403
+from .domain_decomposition_utils import *  # noqa: F403
